personal_chatbot/requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
python-dotenv
pywebview
